modal/requirements.txt
faster-whisper
pyannote.audio
demucs
librosa
pyacoustid
opencv-python-headless
google-cloud-storage
